refactor(settings): extract responses response adapter
