+++ scripts/ci/module-impact.json
@@ -287,7 +287,9 @@
       "ownerPaths": [
         "src/main/settings/backend-resolver.ts",
         "src/main/settings/responses-bridge.ts",
-        "src/main/settings/responses-request-adapter.ts"
+        "src/main/settings/responses-protocol-types.ts",
+        "src/main/settings/responses-request-adapter.ts",
+        "src/main/settings/responses-response-adapter.ts"
       ],
       "interfacePaths": [
         "src/main/settings/backend-resolver.ts",
@@ -304,6 +306,8 @@
           "src/main/settings/backend-resolver.test.ts",
           "src/main/settings/responses-request-adapter.architecture.test.ts",
           "src/main/settings/responses-request-adapter.test.ts",
+          "src/main/settings/responses-response-adapter.architecture.test.ts",
+          "src/main/settings/responses-response-adapter.test.ts",
           "src/main/settings/responses-bridge.test.ts",
           "src/main/settings/responses-bridge.integration.test.ts",
           "src/main/settings/native-responses-compatibility.test.ts",
