+++ src/main/settings/responses-bridge.ts
@@ -8,10 +8,14 @@ import type {
   CustomReasoningEffortTransport,
   ModelReasoningEffort
 } from '../../shared/reasoning-effort'
+import { responsesToChatRequest } from './responses-request-adapter'
 import {
-  responsesToChatRequest,
-  type ResponsesBridgeNamespacedTool
-} from './responses-request-adapter'
+  completionToResponse,
+  ResponsesProtocolError,
+  streamChatToResponses,
+  upstreamErrorMessage
+} from './responses-response-adapter'
+import type { ResponsesBridgeNamespacedTool } from './responses-protocol-types'
 import {
   boundedSkillSelectorCatalog,
   renderSkillSelectorCatalog,
@@ -78,26 +82,6 @@ type ResponsesBridgeOptions = {
 
 type BridgeFetch = typeof fetch
 
-class BridgeHttpError extends Error {
-  constructor(
-    message: string,
-    readonly status: number,
-    readonly type: string
-  ) {
-    super(message)
-    this.name = 'BridgeHttpError'
-  }
-}
-
-const UPSTREAM_IMAGE_TYPES = new Set(['image', 'image_url', 'input_image', 'output_image'])
-
-const unsupportedUpstreamImageOutput = (): BridgeHttpError =>
-  new BridgeHttpError(
-    'Upstream image output is not supported by this gateway',
-    502,
-    'unsupported_upstream_output'
-  )
-
 const json = (response: ServerResponse, status: number, body: unknown): void => {
   response.writeHead(status, { 'content-type': 'application/json' })
   response.end(JSON.stringify(body))
@@ -113,456 +97,6 @@ const readBody = async (request: IncomingMessage): Promise<JsonObject> => {
 // official vendor's exact versioned base, or a custom root normalized to `<root>/v1`), so this only
 // appends `/chat/completions` — preserving any query/hash on the base.
 const chatUrl = (value: string): string => appendChatCompletions(value)
-
-const upstreamErrorMessage = (body: string, status: number): string => {
-  const trimmed = body.trim()
-  if (trimmed) {
-    try {
-      const parsed = JSON.parse(trimmed) as JsonObject
-      const error = parsed.error
-      if (typeof error === 'string') return error
-      if (error && typeof error === 'object' && typeof error.message === 'string') {
-        return error.message
-      }
-      if (typeof parsed.message === 'string') return parsed.message
-    } catch {
-      return trimmed.slice(0, 500)
-    }
-  }
-
-  return `Chat Completions upstream returned ${status}`
-}
-
-const upstreamTextFromContent = (content: unknown): string => {
-  if (content === undefined || content === null) return ''
-  if (typeof content === 'string') return content
-  if (
-    typeof content === 'object' &&
-    UPSTREAM_IMAGE_TYPES.has(String((content as JsonObject).type))
-  ) {
-    throw unsupportedUpstreamImageOutput()
-  }
-  if (!Array.isArray(content)) {
-    throw new BridgeHttpError(
-      'Unsupported upstream message content',
-      502,
-      'unsupported_upstream_output'
-    )
-  }
-
-  return content
-    .map((part) => {
-      if (!part || typeof part !== 'object') {
-        throw new BridgeHttpError(
-          'Unsupported upstream message content part',
-          502,
-          'unsupported_upstream_output'
-        )
-      }
-      if (UPSTREAM_IMAGE_TYPES.has(String(part.type))) throw unsupportedUpstreamImageOutput()
-      if (part.type !== 'text' && part.type !== 'output_text') {
-        throw new BridgeHttpError(
-          `Unsupported upstream message content part: ${String(part.type)}`,
-          502,
-          'unsupported_upstream_output'
-        )
-      }
-      if (typeof part.text !== 'string') {
-        throw new BridgeHttpError(
-          'Upstream text output must contain string text',
-          502,
-          'unsupported_upstream_output'
-        )
-      }
-      return part.text
-    })
-    .join('')
-}
-
-const hasUpstreamImageField = (value: JsonObject): boolean =>
-  (Array.isArray(value.images) && value.images.length > 0) ||
-  value.image !== undefined ||
-  value.image_url !== undefined ||
-  value.output_image !== undefined
-
-const responseFunctionIdentity = (
-  chatName: unknown,
-  tools: readonly ResponsesBridgeNamespacedTool[]
-): { name: string; namespace?: string } => {
-  const name = String(chatName ?? '')
-  const namespaced = tools.find((tool) => `${tool.namespace}__${tool.name}` === name)
-  return namespaced ? { name: namespaced.name, namespace: namespaced.namespace } : { name }
-}
-
-const responseEnvelope = (
-  id: string,
-  model: string,
-  output: JsonObject[],
-  usage?: unknown,
-  status: string = 'completed',
-  error: unknown = null
-): JsonObject => ({
-  id,
-  object: 'response',
-  created_at: Math.floor(Date.now() / 1000),
-  status,
-  error,
-  incomplete_details: null,
-  instructions: null,
-  max_output_tokens: null,
-  model,
-  output,
-  parallel_tool_calls: true,
-  previous_response_id: null,
-  reasoning: { effort: null, summary: null },
-  store: false,
-  temperature: null,
-  text: { format: { type: 'text' } },
-  tool_choice: 'auto',
-  tools: [],
-  top_p: null,
-  truncation: 'disabled',
-  usage: usage ?? null,
-  user: null,
-  metadata: {}
-})
-
-// Chat Completions and Responses use different usage field names. Codex validates the Responses
-// shape before publishing its ACP token-usage update, so passing Chat fields through makes usage
-// silently disappear even though the upstream reported it.
-const chatUsageToResponsesUsage = (usage: unknown): JsonObject | undefined => {
-  if (typeof usage !== 'object' || usage === null || Array.isArray(usage)) return undefined
-
-  const chatUsage = usage as JsonObject
-  const tokenCount = (value: unknown): number | undefined =>
-    typeof value === 'number' && Number.isFinite(value) && value >= 0 ? value : undefined
-  const inputTokens = tokenCount(chatUsage.prompt_tokens)
-  const outputTokens = tokenCount(chatUsage.completion_tokens)
-
-  if (inputTokens === undefined || outputTokens === undefined) return undefined
-
-  const cachedTokens = tokenCount(chatUsage.prompt_tokens_details?.cached_tokens) ?? 0
-  const reasoningTokens = tokenCount(chatUsage.completion_tokens_details?.reasoning_tokens) ?? 0
-
-  return {
-    input_tokens: inputTokens,
-    input_tokens_details: { cached_tokens: cachedTokens },
-    output_tokens: outputTokens,
-    output_tokens_details: { reasoning_tokens: reasoningTokens },
-    total_tokens: tokenCount(chatUsage.total_tokens) ?? inputTokens + outputTokens
-  }
-}
-
-const completionToResponse = (
-  completion: JsonObject,
-  namespacedTools: readonly ResponsesBridgeNamespacedTool[] = []
-): JsonObject => {
-  const message = completion.choices?.[0]?.message ?? {}
-  const output: JsonObject[] = []
-  if (hasUpstreamImageField(message)) throw unsupportedUpstreamImageOutput()
-  // Mirror the streaming path: drop reasoning_content (no faithful Responses representation) and fall
-  // back to a refusal as the visible answer, rather than rejecting model output outright.
-  const contentText = upstreamTextFromContent(message.content)
-  const text =
-    contentText.length > 0
-      ? contentText
-      : typeof message.refusal === 'string' && message.refusal.length > 0
-        ? message.refusal
-        : ''
-  if (text) {
-    output.push({
-      id: `msg_${completion.id}`,
-      type: 'message',
-      status: 'completed',
-      role: 'assistant',
-      content: [{ type: 'output_text', text, annotations: [] }]
-    })
-  }
-  for (const tool of message.tool_calls ?? []) {
-    const identity = responseFunctionIdentity(tool.function?.name, namespacedTools)
-    output.push({
-      id: `fc_${tool.id}`,
-      type: 'function_call',
-      status: 'completed',
-      call_id: tool.id,
-      ...identity,
-      arguments: tool.function?.arguments ?? '{}'
-    })
-  }
-  return responseEnvelope(
-    completion.id ?? `resp_${randomBytes(6).toString('hex')}`,
-    completion.model,
-    output,
-    chatUsageToResponsesUsage(completion.usage)
-  )
-}
-
-const writeEvent = (
-  response: ServerResponse,
-  type: string,
-  sequence: number,
-  fields: JsonObject = {}
-): void => {
-  response.write(
-    `event: ${type}\ndata: ${JSON.stringify({ type, sequence_number: sequence, ...fields })}\n\n`
-  )
-}
-
-const streamChatToResponses = async (
-  upstream: Response,
-  response: ServerResponse,
-  model: string,
-  namespacedTools: readonly ResponsesBridgeNamespacedTool[] = []
-): Promise<{ reasoning: string; callIds: string[] }> => {
-  if (!upstream.body) throw new Error('Chat Completions upstream returned no body')
-  response.writeHead(200, {
-    'content-type': 'text/event-stream',
-    'cache-control': 'no-cache',
-    connection: 'keep-alive'
-  })
-
-  const responseId = `resp_${randomBytes(8).toString('hex')}`
-  const output: JsonObject[] = []
-  const toolItems = new Map<number, { chatId: string; chatName: string; item?: JsonObject }>()
-  let textItem: JsonObject | undefined
-  // Accumulated so the caller can cache it against this turn's tool-call ids (see inputToMessages).
-  let reasoning = ''
-  let usage: JsonObject | undefined
-  let sequence = 0
-  writeEvent(response, 'response.created', sequence++, {
-    response: responseEnvelope(responseId, model, [])
-  })
-  writeEvent(response, 'response.in_progress', sequence++, {
-    response: responseEnvelope(responseId, model, [])
-  })
-
-  const decoder = new TextDecoder()
-  let buffered = ''
-  // Classify how the upstream stream ended so a truncation or token-limit cutoff is never reported as a
-  // clean completion. `terminalFinishReason` is the last finish_reason seen; `sawDone` marks the [DONE]
-  // sentinel. Neither seen ⇒ the connection dropped mid-stream.
-  let terminalFinishReason: string | undefined
-  let sawDone = false
-  const ensureToolItem = (index: number): JsonObject => {
-    const state = toolItems.get(index) ?? { chatId: '', chatName: '' }
-    toolItems.set(index, state)
-    if (state.item) return state.item
-
-    const identity = responseFunctionIdentity(state.chatName, namespacedTools)
-    const callId = state.chatId || `call_${responseId}_${index}`
-    const item: JsonObject = {
-      id: `fc_${callId}_${index}`,
-      type: 'function_call',
-      status: 'in_progress',
-      call_id: callId,
-      ...identity,
-      arguments: ''
-    }
-    state.item = item
-    output.push(item)
-    writeEvent(response, 'response.output_item.added', sequence++, {
-      output_index: output.indexOf(item),
-      item
-    })
-    return item
-  }
-  const consume = (chunk: JsonObject): void => {
-    usage = chatUsageToResponsesUsage(chunk.usage) ?? usage
-    const finishReason = chunk.choices?.[0]?.finish_reason
-    if (typeof finishReason === 'string' && finishReason.length > 0) {
-      terminalFinishReason = finishReason
-    }
-    const delta = chunk.choices?.[0]?.delta ?? {}
-    if (hasUpstreamImageField(delta)) throw unsupportedUpstreamImageOutput()
-    // Never throw on model output mid-stream: the turn's headers are already sent, so a throw would
-    // reset the socket and reach the agent as an opaque "error decoding response body". Reasoning-model
-    // providers stream `reasoning_content` deltas that have no faithful Responses representation here,
-    // so drop them; a `refusal` IS the model's answer, so surface it as visible text.
-    if (typeof delta.reasoning_content === 'string') reasoning += delta.reasoning_content
-    const contentText = upstreamTextFromContent(delta.content)
-    const textDelta =
-      contentText.length > 0
-        ? contentText
-        : typeof delta.refusal === 'string' && delta.refusal.length > 0
-          ? delta.refusal
-          : ''
-    if (textDelta) {
-      if (!textItem) {
-        textItem = {
-          id: `msg_${responseId}`,
-          type: 'message',
-          status: 'in_progress',
-          role: 'assistant',
-          content: []
-        }
-        output.push(textItem)
-        writeEvent(response, 'response.output_item.added', sequence++, {
-          output_index: output.length - 1,
-          item: textItem
-        })
-        writeEvent(response, 'response.content_part.added', sequence++, {
-          item_id: textItem.id,
-          output_index: output.length - 1,
-          content_index: 0,
-          part: { type: 'output_text', text: '', annotations: [] }
-        })
-      }
-      writeEvent(response, 'response.output_text.delta', sequence++, {
-        item_id: textItem.id,
-        output_index: output.indexOf(textItem),
-        content_index: 0,
-        delta: textDelta
-      })
-      textItem.content.push({ type: 'output_text', text: textDelta, annotations: [] })
-    }
-    for (const call of delta.tool_calls ?? []) {
-      const index = Number(call.index ?? 0)
-      const state = toolItems.get(index) ?? { chatId: '', chatName: '' }
-      toolItems.set(index, state)
-      if (typeof call.id === 'string') state.chatId += call.id
-      if (typeof call.function?.name === 'string') state.chatName += call.function.name
-      const argumentsDelta = call.function?.arguments ?? ''
-      if (!argumentsDelta && !state.item) continue
-      const item = ensureToolItem(index)
-      item.arguments += argumentsDelta
-      if (argumentsDelta) {
-        writeEvent(response, 'response.function_call_arguments.delta', sequence++, {
-          item_id: item.id,
-          output_index: output.indexOf(item),
-          delta: argumentsDelta
-        })
-      }
-    }
-  }
-
-  const handleRecord = (record: string): void => {
-    const data = record
-      .split(/\r?\n/)
-      .filter((line) => line.startsWith('data:'))
-      .map((line) => line.slice(5).trim())
-      .join('\n')
-    if (data === '[DONE]') sawDone = true
-    else if (data) consume(JSON.parse(data) as JsonObject)
-  }
-
-  // A mid-stream read error (socket reset, fetch timeout) throws here. Headers are already sent, so
-  // instead of letting it bubble to the server's catch — which would abruptly destroy the socket and
-  // surface as an opaque "error decoding response body" — record it as a truncation and fall through
-  // to emit a terminal response.failed below.
-  let streamError: unknown
-  try {
-    for await (const chunk of upstream.body) {
-      buffered += decoder.decode(chunk, { stream: true })
-      // SSE records are separated by a blank line; tolerate both LF and CRLF framing.
-      const records = buffered.split(/\r?\n\r?\n/)
-      buffered = records.pop() ?? ''
-      for (const record of records) handleRecord(record)
-    }
-    // Flush a trailing record not terminated by a blank line (e.g. a final `data: [DONE]`) so its
-    // terminal signal is not lost.
-    buffered += decoder.decode()
-    if (buffered.trim()) handleRecord(buffered)
-  } catch (error) {
-    streamError = error
-  }
-
-  // A valid no-argument tool call may never stream an arguments delta, so materialize any buffered
-  // id/name pair before completing output items.
-  for (const index of toolItems.keys()) ensureToolItem(index)
-
-  for (const item of output) {
-    item.status = 'completed'
-    const outputIndex = output.indexOf(item)
-    if (item.type === 'message') {
-      const text = item.content.map((part: JsonObject) => part.text).join('')
-      item.content = [{ type: 'output_text', text, annotations: [] }]
-      writeEvent(response, 'response.output_text.done', sequence++, {
-        item_id: item.id,
-        output_index: outputIndex,
-        content_index: 0,
-        text
-      })
-      writeEvent(response, 'response.content_part.done', sequence++, {
-        item_id: item.id,
-        output_index: outputIndex,
-        content_index: 0,
-        part: item.content[0]
-      })
-    } else {
-      writeEvent(response, 'response.function_call_arguments.done', sequence++, {
-        item_id: item.id,
-        output_index: outputIndex,
-        arguments: item.arguments
-      })
-    }
-    writeEvent(response, 'response.output_item.done', sequence++, {
-      output_index: outputIndex,
-      item
-    })
-  }
-  // Classification priority: an explicit clean finish_reason wins (the answer finished even if trailing
-  // bytes then errored); otherwise a mid-stream error or a non-terminal finish_reason (`length`,
-  // `content_filter`) is a cut-off answer; a bare [DONE] with no finish_reason still counts as a proper
-  // termination; anything else means the stream dropped with no terminal signal at all.
-  if (streamError instanceof BridgeHttpError) {
-    log.warn('bridge unsupported upstream output', { model, type: streamError.type })
-    writeEvent(response, 'response.failed', sequence++, {
-      response: responseEnvelope(responseId, model, output, undefined, 'failed', {
-        type: streamError.type,
-        message: streamError.message
-      })
-    })
-  } else if (terminalFinishReason === 'stop' || terminalFinishReason === 'tool_calls') {
-    writeEvent(response, 'response.completed', sequence++, {
-      response: responseEnvelope(responseId, model, output, usage)
-    })
-  } else if (streamError) {
-    log.warn('bridge stream error', {
-      model,
-      error: streamError instanceof Error ? streamError.message : String(streamError)
-    })
-    writeEvent(response, 'response.failed', sequence++, {
-      response: responseEnvelope(responseId, model, output, usage, 'failed', {
-        type: 'upstream_error',
-        message: 'Upstream stream ended before completion'
-      })
-    })
-  } else if (terminalFinishReason) {
-    // A non-terminal finish_reason (e.g. `length`, `content_filter`) is a truncated answer, not a
-    // complete one — surface it as incomplete so the agent doesn't treat a cut-off as a full result.
-    log.warn('bridge stream incomplete', { model, finishReason: terminalFinishReason })
-    writeEvent(response, 'response.incomplete', sequence++, {
-      response: {
-        ...responseEnvelope(responseId, model, output, usage, 'incomplete'),
-        incomplete_details: { reason: terminalFinishReason }
-      }
-    })
-  } else if (sawDone) {
-    writeEvent(response, 'response.completed', sequence++, {
-      response: responseEnvelope(responseId, model, output, usage)
-    })
-  } else {
-    // No finish_reason and no [DONE]: the upstream ended mid-stream without a terminal signal.
-    log.warn('bridge stream truncated (no terminal finish_reason)', { model })
-    writeEvent(response, 'response.failed', sequence++, {
-      response: responseEnvelope(responseId, model, output, usage, 'failed', {
-        type: 'upstream_incomplete',
-        message: 'Upstream stream ended without a terminal finish_reason'
-      })
-    })
-  }
-  response.end()
-
-  const toolCalls = output.filter((item) => item.type === 'function_call')
-  log.info('bridge turn completed (stream)', {
-    model,
-    textItems: output.filter((item) => item.type === 'message').length,
-    toolCalls: toolCalls.length,
-    toolNames: toolCalls.map((item) => item.name)
-  })
-
-  return { reasoning, callIds: toolCalls.map((item) => String(item.call_id)) }
-}
 
 export class ResponsesBridge {
   private server: Server | undefined
@@ -746,7 +280,7 @@ export class ResponsesBridge {
       void this.handle(request, response).catch((error: unknown) => {
         if (response.destroyed || response.writableEnded) return
         if (!response.headersSent) {
-          const bridgeError = error instanceof BridgeHttpError ? error : undefined
+          const bridgeError = error instanceof ResponsesProtocolError ? error : undefined
           json(response, bridgeError?.status ?? 400, {
             error: {
               type: bridgeError?.type ?? 'invalid_request_error',
@@ -942,9 +476,5 @@ export class ResponsesBridge {
 }
 
 export { chatUrl, completionToResponse, upstreamErrorMessage }
-export {
-  inputToMessages,
-  responsesToChatRequest,
-  toolsToChat,
-  type ResponsesBridgeNamespacedTool
-} from './responses-request-adapter'
+export { inputToMessages, responsesToChatRequest, toolsToChat } from './responses-request-adapter'
+export type { ResponsesBridgeNamespacedTool } from './responses-protocol-types'
